pyproject.toml
[tool.poetry]
name = "orguns"
version = "0.1.0"
description = ""
authors = ["mohseneptune <mohseneptune@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.25"
psycopg = "^3.1.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
black = "^23.12.1"
pre-commit = "^3.6.0"
alembic = "^1.13.1"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
